refactor: unify HEADER_BITS type and simplify calculations
- Changed the type of `HEADER_BITS` from `u8` to `usize` for consistency across the codebase.
- Simplified calculations involving `HEADER_BITS` in `extract_text`, `embed_text`, and `max_message_size` functions to remove unnecessary type conversions.

+++ src/stego/decode.rs
@@ -36,7 +36,7 @@ use super::{
 pub fn extract_text(image: &RgbImage) -> Result<String, StegoError>
 {
     let available_bits = channel_capacity_bits(image);
-    if available_bits < usize::from(HEADER_BITS)
+    if available_bits < HEADER_BITS
     {
         return Err(StegoError::MissingHeader { available_bits });
     }
@@ -72,7 +72,7 @@ pub fn extract_text(image: &RgbImage) -> Result<String, StegoError>
 
     // Our data starts after the header, so we need to subtract the header
     let remaining_capacity_bytes = {
-        let remaining_capacity_bits = available_bits - usize::from(HEADER_BITS);
+        let remaining_capacity_bits = available_bits - HEADER_BITS;
         remaining_capacity_bits / 8
     };
